\name{contours}\alias{contours}\docType{data}\title{An \code{SpatialPolygonsDataFrame} corresponding to the limits of the landscape r2
}\usage{data(contours)}\description{ contours is a \code{SpatialPolygonsDataFrame} object corresponding to the limits of the landscape provided depicted is object r2.
}\format{ An object of class \code{SpatialPolygonsDataFrame} (see package \code{sp}).
}
 \source{The original data were made available by IGN (http://www.ign.fr/) and further rasterized using QGIS (http://www.qgis.org/fr/site/).
 } \examples{
 data(r2)
}