\name{DONBIN}
\alias{DONBIN}
\docType{data}
\title{A \code{data.frame} object corresponding to 319 rows and 9 variables.}
\usage{
data(DONBIN)
}
\description{ A \code{data.frame} object corresponding to 319 rows and 9 variables.}
\format{
 An object of class \code{data.frame}).
}
 \examples{
data(DONBIN)
dim(DONBIN)
}

