\name{DONPG}
\alias{DONPG}
\docType{data}
\title{A \code{data.frame} object corresponding to 319 rows and 7 variables.}
\usage{
data(DONPG)
}
\description{ A \code{data.frame} object corresponding to 319 rows and 7 variables.}
\format{
 An object of class \code{data.frame}).
}
 \examples{
data(DONPG)
dim(DONPG)
}

