\name{ECPaysage-package}
\alias{ECPaysage-package}
\alias{ECPaysage}
\docType{package}
\title{ A set of R resources to teach landscape ecology, landscape analysis and spatial statistics}
\description{ The package provides some functions useful for teaching landscape analysis and spatial statistics. ECPaysage also provides a handful of useful datasets. A good place to start is the vignette, which you can access by typing \code{vignette("introduction_ecpaysage")}.}
\details{ See package vignette}
\author{Jean-Pierre Rossi, INRA}
\examples{vignette("introduction_ecpaysage")}