\name{ecpaysage2-package}
\alias{ecpaysage2-package}
\alias{ecpaysage2}
\docType{package}
\title{ A set of R resources to teach landscape ecology, landscape analysis and spatial statistics
}
\description{ The package provides a handful of datasets and some information useful for teaching landscape analysis and spatial statistics.
}
\details{See package vignette}
\author{Jean-Pierre Rossi, INRA}

\examples{vignette("introduction_ecpaysage")}


