\name{l104pt}\alias{l104pt}\docType{data}\title{A list containing 104 \code{rasterLayer} objects depicting the local landscape located at 104 sampling points regularly distributed
 in north of Montpellier (France)
}\usage{data(l104pt)}\description{ The list contains the buffers extracted at each of 104 sampling points regularly distributed. The landscape is the one given in object \code{r2}. The buffers are circular with radius of 600 m. The grain is 50 m.
}\format{ A list of objects of class \code{RasterLayer} (see package \code{raster}).
}
 \source{The original data were made available by IGN (http://www.ign.fr/) and further rasterized using QGIS (http://www.qgis.org/fr/site/).
 } \examples{
data(l104pt)

}