\name{r1}\alias{r1}\docType{data}\title{An example of \code{rasterLayer} object
}\usage{data(r1)}\description{ r1 is a 100 by 100 \code{RasterLayer} object. The data have been simulated so as to exhibit a clear spatial structure.
}\format{ An object of class \code{RasterLayer} (see package \code{raster}).
} \examples{
data(r1)
}