\name{r2}\alias{r2}\docType{data}\title{An \code{rasterLayer} object depicting landuses in north of Montpellier (France)
}\usage{data(r2)}\description{ r2 is a 528 by 841 \code{RasterLayer} object depicting landuses in north of Montpellier (France). The grain is 50 m.
}\format{ An object of class \code{RasterLayer} (see package \code{raster}).
}
 \source{The original data were made available by IGN (http://www.ign.fr/) and further rasterized using QGIS (http://www.qgis.org/fr/site/).
 } \examples{
 data(r2)

}