\name{xy104pt}\alias{xy104pt}\docType{data}\title{A \code{SpatialPoints} object corresponding to 104 psampling points regularly distributed across a real landscape in north of Montpellier (France)
}\usage{data(xy104pt)}\description{ The \code{SpatialPoints} were generated using the function \code{spsample}. 
}\format{ An object of class \code{SpatialPoints} (see package \code{sp}).
}
 \examples{
data(xy104pt)


}